# docs/openapi-frontend.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/guilds": {
      "get": {
        "operationId": "GuildsController_getGuilds",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Guilds"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/guilds/onboard": {
      "post": {
        "operationId": "GuildsController_onboardGuild",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OnboardGuildDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Guilds"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/guilds/{guildId}/settings": {
      "get": {
        "operationId": "GuildsController_getSettings",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Guilds"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "GuildsController_updateSettings",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchSettingsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Guilds"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/guilds/{guildId}/modules": {
      "patch": {
        "operationId": "GuildsController_updateModules",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchGuildModulesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Guilds"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "GuildsController_getModules",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Guilds"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/guilds/{guildId}/token": {
      "patch": {
        "operationId": "GuildsController_updateToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchGuildTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Guilds"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/guilds/{guildId}/channels": {
      "get": {
        "operationId": "GuildsController_getChannels",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Guilds"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/guilds/{guildId}/stats": {
      "get": {
        "operationId": "GuildsController_getStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Guilds"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/guilds/{guildId}/bot-status": {
      "get": {
        "operationId": "GuildsController_getBotStatus",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Guilds"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/guilds/{guildId}/activity-sparkline": {
      "get": {
        "operationId": "GuildsController_getActivitySparkline",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Guilds"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/guilds/{guildId}/logs/settings": {
      "get": {
        "operationId": "LogsController_getSettings",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Logs"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "LogsController_patchSettings",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchLogSettingsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Logs"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/guilds/{guildId}/logs/events": {
      "get": {
        "operationId": "LogsController_getEvents",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Logs"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Frontend API",
    "description": "Public API. Authentication: Bearer JWT in Authorization header. All requests/responses use application/json. Envelope pattern: success in data, errors in error.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "OnboardGuildDto": {
        "type": "object",
        "properties": {}
      },
      "PatchSettingsDto": {
        "type": "object",
        "properties": {}
      },
      "PatchGuildModulesDto": {
        "type": "object",
        "properties": {}
      },
      "PatchGuildTokenDto": {
        "type": "object",
        "properties": {}
      },
...
      "PatchLogSettingsDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}